// src/app/Shared/models/account.interface.ts
export interface IAccount {
    accountName: string;
    accountNo: string;
    accountType: string;
    authBy: string;
    bankAdd: string;
    bankBranch: string;
    bankCity: string;
    bankName: string;
    bankNo: string;
    bankSwiftCd: string;
    bankTown: string;
    blocked: boolean;
    blockedDate: string;
    clientNo: string;
    country: string;
    currency: string;
    no: string;
    payeeNo: string;
    sortCode: string;
  } 